type is switcghed for tinyint

+++ database/migrations/2020_08_09_071033_add_field_on_accounts_table.php
@@ -14,7 +14,7 @@ class AddFieldOnAccountsTable extends Migration
     public function up()
     {
         Schema::table('accounts', function (Blueprint $table) {
-            $table->unsignedInteger('account_type');
+            $table->unsignedTinyInteger('account_type');
         });
     }
